requirements.txt
pydantic
pydantic_settings
pypubsub
